+++ madmom/features/beats.py
@@ -143,7 +143,7 @@ class MultiModelSelectionProcessor(Processor):
     Examples
     --------
     The MultiModelSelectionProcessor takes a list of model predictions as it's
-    call argument. Thus, `ppost_processor` of `RNNBeatProcessor` hast to be set
+    call argument. Thus, `post_processor` of `RNNBeatProcessor` hast to be set
     to 'None' in order to get the predictions of all models.
 
     >>> proc = RNNBeatProcessor(post_processor=None)
@@ -771,6 +771,42 @@ class CRFBeatDetectionProcessor(BeatTrackingProcessor):
         return g
 
 
+def threshold_activations(activations, threshold):
+    """
+    Threshold activations to include only the main segment exceeding the given
+    threshold (i.e. first to last time/index exceeding the threshold).
+
+    Parameters
+    ----------
+    activations : numpy array
+        Activations to be thresholded.
+    threshold : float
+        Threshold value.
+
+    Returns
+    -------
+    activations : numpy array
+        Thresholded activations
+    start : int
+        Index of the first activation exceeding the threshold.
+
+    Notes
+    -----
+
+    This function can be used to extract the main segment of beat activations
+    to track only the beats where the activations exceed the threshold.
+
+    """
+    first = last = 0
+    # use only the activations > threshold
+    idx = np.nonzero(activations >= threshold)[0]
+    if idx.any():
+        first = max(first, np.min(idx))
+        last = min(len(activations), np.max(idx) + 1)
+    # return thresholded activations segment and first index
+    return activations[first:last], first
+
+
 class DBNBeatTrackingProcessor(OnlineProcessor):
     """
     Beat tracking with RNNs and a dynamic Bayesian network (DBN) approximated
@@ -869,7 +905,7 @@ class DBNBeatTrackingProcessor(OnlineProcessor):
         self.fps = fps
         self.min_bpm = min_bpm
         self.max_bpm = max_bpm
-        # kepp state in online mode
+        # keep state in online mode
         self.online = online
         # TODO: refactor the visualisation stuff
         if self.online:
@@ -913,18 +949,16 @@ class DBNBeatTrackingProcessor(OnlineProcessor):
         first = 0
         # use only the activations > threshold
         if self.threshold:
-            idx = np.nonzero(activations >= self.threshold)[0]
-            if idx.any():
-                first = max(first, np.min(idx))
-                last = min(len(activations), np.max(idx) + 1)
-            else:
-                last = first
-            activations = activations[first:last]
-        # return the beats if no activations given / remain after thresholding
+            activations, first = threshold_activations(activations,
+                                                       self.threshold)
+        # return no beats if no activations given / remain after thresholding
         if not activations.any():
             return beats
         # get the best state path by calling the viterbi algorithm
         path, _ = self.hmm.viterbi(activations)
+        # also return no beats if no path was found
+        if not path.any():
+            return beats
         # correct the beat positions if needed
         if self.correct:
             # for each detection determine the "beat range", i.e. states where