+++ madmom/io/audio.py
@@ -14,6 +14,7 @@ import os
 import subprocess
 import sys
 import tempfile
+import io
 
 import numpy as np
 
@@ -70,7 +71,8 @@ def _ffmpeg_fmt(dtype):
 
 
 def _ffmpeg_call(infile, output, fmt='f32le', sample_rate=None, num_channels=1,
-                 skip=None, max_len=None, cmd='ffmpeg'):
+                 channel=None, skip=None, max_len=None, cmd='ffmpeg',
+                 replaygain_mode=None, replaygain_preamp=0.0):
     """
     Create a sequence of strings indicating ffmpeg how to be called as well as
     the parameters necessary to decode the given input (file) to the given
@@ -90,12 +92,21 @@ def _ffmpeg_call(infile, output, fmt='f32le', sample_rate=None, num_channels=1,
         Sample rate to re-sample the signal to (if set) [Hz].
     num_channels : int, optional
         Number of channels to reduce the signal to.
+        If 'None', return the signal with its original channels,
+        or whatever is selected by `channel`.
+    channel : int, optional
+        When reducing a signal to `num_channels` of 1, use this channel,
+        or 'None' to return the average across all channels.
     skip : float, optional
         Number of seconds to skip at beginning of file.
     max_len : float, optional
         Maximum length in seconds to decode.
     cmd : {'ffmpeg','avconv'}, optional
         Decoding command (defaults to ffmpeg, alternatively supports avconv).
+    replaygain_mode : {None, 'track','album'}, optional
+        Specify the ReplayGain volume-levelling mode (None to disable).
+    replaygain_preamp : float, optional
+        ReplayGain preamp volume change level (in dB).
 
     Returns
     -------
@@ -121,6 +132,8 @@ def _ffmpeg_call(infile, output, fmt='f32le', sample_rate=None, num_channels=1,
         in_ac = str(int(infile.num_channels))
         in_ar = str(int(infile.sample_rate))
         infile = str("pipe:0")
+    elif isinstance(infile, io.IOBase):
+        infile = "-"
     else:
         infile = str(infile)
     # general options
@@ -133,6 +146,10 @@ def _ffmpeg_call(infile, output, fmt='f32le', sample_rate=None, num_channels=1,
     if infile == "pipe:0":
         call.extend(["-f", in_fmt, "-ac", in_ac, "-ar", in_ar])
     call.extend(["-i", infile])
+    if replaygain_mode:
+        audio_filter = ("volume=replaygain=%s:replaygain_preamp=%.1f"
+                        % (replaygain_mode, replaygain_preamp))
+        call.extend(["-af", audio_filter])
     # output options
     call.extend(["-f", str(fmt)])
     if max_len:
@@ -141,6 +158,9 @@ def _ffmpeg_call(infile, output, fmt='f32le', sample_rate=None, num_channels=1,
     # output options
     if num_channels:
         call.extend(["-ac", str(int(num_channels))])
+    if channel is not None and (num_channels == 1 or num_channels is None):
+        # Calling with channel=x and num_channels
+        call.extend(["-af", "pan=mono|c0=c%d" % int(channel)])
     if sample_rate:
         call.extend(["-ar", str(int(sample_rate))])
     call.append(output)
@@ -148,8 +168,9 @@ def _ffmpeg_call(infile, output, fmt='f32le', sample_rate=None, num_channels=1,
 
 
 def decode_to_disk(infile, fmt='f32le', sample_rate=None, num_channels=1,
-                   skip=None, max_len=None, outfile=None, tmp_dir=None,
-                   tmp_suffix=None, cmd='ffmpeg'):
+                   channel=None, skip=None, max_len=None, outfile=None,
+                   tmp_dir=None, tmp_suffix=None, cmd='ffmpeg',
+                   replaygain_mode=None, replaygain_preamp=0.0):
     """
     Decode the given audio file to another file.
 
@@ -165,6 +186,11 @@ def decode_to_disk(infile, fmt='f32le', sample_rate=None, num_channels=1,
         Sample rate to re-sample the signal to (if set) [Hz].
     num_channels : int, optional
         Number of channels to reduce the signal to.
+        If 'None', return the signal with its original channels,
+        or whatever is selected by `channel`.
+    channel : int, optional
+        When reducing a signal to `num_channels` of 1, use this channel,
+        or 'None' to return the average across all channels.
     skip : float, optional
         Number of seconds to skip at beginning of file.
     max_len : float, optional
@@ -180,6 +206,10 @@ def decode_to_disk(infile, fmt='f32le', sample_rate=None, num_channels=1,
         ".pcm" (including the dot).
     cmd : {'ffmpeg', 'avconv'}, optional
         Decoding command (defaults to ffmpeg, alternatively supports avconv).
+    replaygain_mode : {None, 'track','album'}, optional
+        Specify the ReplayGain volume-levelling mode (None to disable).
+    replaygain_preamp : float, optional
+        ReplayGain preamp volume change level (in dB).
 
     Returns
     -------
@@ -208,7 +238,9 @@ def decode_to_disk(infile, fmt='f32le', sample_rate=None, num_channels=1,
     # call ffmpeg (throws exception on error)
     try:
         call = _ffmpeg_call(infile, outfile, fmt, sample_rate, num_channels,
-                            skip, max_len, cmd)
+                            channel, skip, max_len, cmd,
+                            replaygain_mode=replaygain_mode,
+                            replaygain_preamp=replaygain_preamp)
         subprocess.check_call(call)
     except Exception:
         if delete_on_fail:
@@ -218,7 +250,8 @@ def decode_to_disk(infile, fmt='f32le', sample_rate=None, num_channels=1,
 
 
 def decode_to_pipe(infile, fmt='f32le', sample_rate=None, num_channels=1,
-                   skip=None, max_len=None, buf_size=-1, cmd='ffmpeg'):
+                   channel=None, skip=None, max_len=None, buf_size=-1,
+                   cmd='ffmpeg', replaygain_mode=None, replaygain_preamp=0.0):
     """
     Decode the given audio and return a file-like object for reading the
     samples, as well as a process object.
@@ -235,6 +268,11 @@ def decode_to_pipe(infile, fmt='f32le', sample_rate=None, num_channels=1,
         Sample rate to re-sample the signal to (if set) [Hz].
     num_channels : int, optional
         Number of channels to reduce the signal to.
+        If 'None', return the signal with its original channels,
+        or whatever is selected by `channel`.
+    channel : int, optional
+        When reducing a signal to `num_channels` of 1, use this channel,
+        or 'None' to return the average across all channels.
     skip : float, optional
         Number of seconds to skip at beginning of file.
     max_len : float, optional
@@ -246,6 +284,10 @@ def decode_to_pipe(infile, fmt='f32le', sample_rate=None, num_channels=1,
         - '1' means line-buffered, any other value is the buffer size in bytes.
     cmd : {'ffmpeg','avconv'}, optional
         Decoding command (defaults to ffmpeg, alternatively supports avconv).
+    replaygain_mode : {None, 'track','album'}, optional
+        Specify the ReplayGain volume-levelling mode (None to disable).
+    replaygain_preamp : float, optional
+        ReplayGain preamp volume change level (in dB).
 
     Returns
     -------
@@ -261,17 +303,19 @@ def decode_to_pipe(infile, fmt='f32le', sample_rate=None, num_channels=1,
 
     """
     # check input file type
-    if not isinstance(infile, (string_types, Signal)):
-        raise ValueError("only file names or Signal instances are supported "
-                         "as `infile`, not %s." % infile)
+    if not isinstance(infile, (string_types, io.IOBase, Signal)):
+        raise ValueError("only file names, file objects or Signal instances "
+                         "are supported as `infile`, not %s." % infile)
     # Note: closing the file-like object only stops decoding because ffmpeg
     #       reacts on that. A cleaner solution would be calling proc.terminate
     #       explicitly, but this is only available in Python 2.6+. proc.wait
     #       needs to be called in any case.
-    call = _ffmpeg_call(infile, "pipe:1", fmt, sample_rate, num_channels, skip,
-                        max_len, cmd)
+    call = _ffmpeg_call(infile, "pipe:1", fmt, sample_rate, num_channels,
+                        channel, skip, max_len, cmd,
+                        replaygain_mode=replaygain_mode,
+                        replaygain_preamp=replaygain_preamp)
     # redirect stdout to a pipe and buffer as requested
-    if isinstance(infile, Signal):
+    if isinstance(infile, (Signal, io.IOBase)):
         proc = subprocess.Popen(call, stdin=subprocess.PIPE,
                                 stdout=subprocess.PIPE, bufsize=buf_size)
     else:
@@ -280,7 +324,8 @@ def decode_to_pipe(infile, fmt='f32le', sample_rate=None, num_channels=1,
 
 
 def decode_to_memory(infile, fmt='f32le', sample_rate=None, num_channels=1,
-                     skip=None, max_len=None, cmd='ffmpeg'):
+                     channel=None, skip=None, max_len=None, cmd='ffmpeg',
+                     replaygain_mode=None, replaygain_preamp=0.0):
     """
     Decode the given audio and return it as a binary string representation.
 
@@ -296,12 +341,21 @@ def decode_to_memory(infile, fmt='f32le', sample_rate=None, num_channels=1,
         Sample rate to re-sample the signal to (if set) [Hz].
     num_channels : int, optional
         Number of channels to reduce the signal to.
+        If 'None', return the signal with its original channels,
+        or whatever is selected by `channel`.
+    channel : int, optional
+        When reducing a signal to `num_channels` of 1, use this channel,
+        or 'None' to return the average across all channels.
     skip : float, optional
         Number of seconds to skip at beginning of file.
     max_len : float, optional
         Maximum length in seconds to decode.
     cmd : {'ffmpeg', 'avconv'}, optional
         Decoding command (defaults to ffmpeg, alternatively supports avconv).
+    replaygain_mode : {None, 'track','album'}, optional
+        Specify the ReplayGain volume-levelling mode (None to disable).
+    replaygain_preamp : float, optional
+        ReplayGain preamp volume change level (in dB).
 
     Returns
     -------
@@ -310,13 +364,15 @@ def decode_to_memory(infile, fmt='f32le', sample_rate=None, num_channels=1,
 
     """
     # check input file type
-    if not isinstance(infile, (string_types, Signal)):
-        raise ValueError("only file names or Signal instances are supported "
-                         "as `infile`, not %s." % infile)
+    if not isinstance(infile, (string_types, io.IOBase, Signal)):
+        raise ValueError("only file names, file objects or Signal instances "
+                         "are supported as `infile`, not %s." % infile)
     # prepare decoding to pipe
     _, proc = decode_to_pipe(infile, fmt=fmt, sample_rate=sample_rate,
-                             num_channels=num_channels, skip=skip,
-                             max_len=max_len, cmd=cmd)
+                             num_channels=num_channels, channel=channel,
+                             skip=skip, max_len=max_len, cmd=cmd,
+                             replaygain_mode=replaygain_mode,
+                             replaygain_preamp=replaygain_preamp)
     # decode the input to memory
     if isinstance(infile, Signal):
         # Note: np.getbuffer was removed in Python 3, but Python 2 memoryviews
@@ -326,6 +382,9 @@ def decode_to_memory(infile, fmt='f32le', sample_rate=None, num_channels=1,
         except AttributeError:
             mv = memoryview(infile)
             signal, _ = proc.communicate(mv.cast('b'))
+    elif isinstance(infile, io.IOBase):
+        signal, _ = proc.communicate(infile.read())
+        infile.seek(0)
     else:
         signal, _ = proc.communicate()
     if proc.returncode != 0:
@@ -357,8 +416,19 @@ def get_file_info(infile, cmd='ffprobe'):
         info['sample_rate'] = infile.sample_rate
     else:
         # call ffprobe
-        output = subprocess.check_output([cmd, "-v", "quiet", "-show_streams",
-                                          infile])
+        if isinstance(infile, io.IOBase):
+            call = [cmd, "-v", "quiet", "-show_streams", "pipe:0"]
+            proc = subprocess.Popen(call, stdin=subprocess.PIPE,
+                                    stdout=subprocess.PIPE)
+            output, _ = proc.communicate(infile.read())
+            retcode = proc.poll()
+            infile.seek(0)
+            if retcode:
+                raise subprocess.CalledProcessError(retcode, call,
+                                                    output=output)
+        else:
+            output = subprocess.check_output([cmd, "-v", "quiet",
+                                              "-show_streams", infile])
         # parse information
         for line in output.split():
             if line.startswith(b'channels='):
@@ -373,8 +443,9 @@ def get_file_info(infile, cmd='ffprobe'):
 
 
 def load_ffmpeg_file(filename, sample_rate=None, num_channels=None,
-                     start=None, stop=None, dtype=None,
-                     cmd_decode='ffmpeg', cmd_probe='ffprobe'):
+                     channel=None, start=None, stop=None, dtype=None,
+                     cmd_decode='ffmpeg', cmd_probe='ffprobe',
+                     replaygain_mode=None, replaygain_preamp=0.0):
     """
     Load the audio data from the given file and return it as a numpy array.
 
@@ -390,8 +461,12 @@ def load_ffmpeg_file(filename, sample_rate=None, num_channels=None,
         Sample rate to re-sample the signal to [Hz]; 'None' returns the signal
         in its original rate.
     num_channels : int, optional
-        Reduce or expand the signal to `num_channels` channels; 'None' returns
-        the signal with its original channels.
+        Reduce or expand the signal to `num_channels` channels.
+        If 'None', return the signal with its original channels,
+        or whatever is selected by `channel`.
+    channel : int, optional
+        When reducing a signal to `num_channels` of 1, use this channel,
+        or 'None' to return the average across all channels.
     start : float, optional
         Start position [seconds].
     stop : float, optional
@@ -404,6 +479,10 @@ def load_ffmpeg_file(filename, sample_rate=None, num_channels=None,
         Decoding command (defaults to ffmpeg, alternatively supports avconv).
     cmd_probe : {'ffprobe', 'avprobe'}, optional
         Probing command (defaults to ffprobe, alternatively supports avprobe).
+    replaygain_mode : {None, 'track','album'}, optional
+        Specify the ReplayGain volume-levelling mode (None to disable).
+    replaygain_preamp : float, optional
+        ReplayGain preamp volume change level (in dB).
 
     Returns
     -------
@@ -428,8 +507,12 @@ def load_ffmpeg_file(filename, sample_rate=None, num_channels=None,
     signal = np.frombuffer(decode_to_memory(filename, fmt=fmt,
                                             sample_rate=sample_rate,
                                             num_channels=num_channels,
+                                            channel=channel,
                                             skip=start, max_len=max_len,
-                                            cmd=cmd_decode),
+                                            cmd=cmd_decode,
+                                            replaygain_mode=replaygain_mode,
+                                            replaygain_preamp=replaygain_preamp
+                                            ),
                            dtype=dtype)
     # get the needed information from the file
     if sample_rate is None or num_channels is None:
@@ -445,8 +528,8 @@ def load_ffmpeg_file(filename, sample_rate=None, num_channels=None,
 
 
 # functions for loading/saving wave files
-def load_wave_file(filename, sample_rate=None, num_channels=None, start=None,
-                   stop=None, dtype=None):
+def load_wave_file(filename, sample_rate=None, num_channels=None, channel=None,
+                   start=None, stop=None, dtype=None):
     """
     Load the audio data from the given file and return it as a numpy array.
 
@@ -462,8 +545,12 @@ def load_wave_file(filename, sample_rate=None, num_channels=None, start=None,
         Desired sample rate of the signal [Hz], or 'None' to return the
         signal in its original rate.
     num_channels : int, optional
-        Reduce or expand the signal to `num_channels` channels, or 'None'
-        to return the signal with its original channels.
+        Reduce or expand the signal to `num_channels` channels
+        If 'None', return the signal with its original channels,
+        or whichever is selected by `channel`.
+    channel : int, optional
+        When reducing a signal to `num_channels` of 1, use this channel,
+        or 'None' to return the average across all channels.
     start : float, optional
         Start position [seconds].
     stop : float, optional
@@ -507,10 +594,12 @@ def load_wave_file(filename, sample_rate=None, num_channels=None, start=None,
         stop = min(len(signal), int(stop * file_sample_rate))
     if start is not None or stop is not None:
         signal = signal[start: stop]
-    # up-/down-mix if needed
+    if channel is not None and num_channels is None:
+        # It's clear what the caller means here
+        num_channels = 1
     if num_channels is not None:
         from ..audio.signal import remix
-        signal = remix(signal, num_channels)
+        signal = remix(signal, num_channels, channel)
     # return the signal
     return signal, file_sample_rate
 
@@ -548,8 +637,9 @@ def write_wave_file(signal, filename, sample_rate=None):
 
 
 # function for automatically determining how to open audio files
-def load_audio_file(filename, sample_rate=None, num_channels=None, start=None,
-                    stop=None, dtype=None):
+def load_audio_file(filename, sample_rate=None, num_channels=None,
+                    channel=None, start=None, stop=None, dtype=None,
+                    replaygain_mode=None, replaygain_preamp=0.0):
     """
     Load the audio data from the given file and return it as a numpy array.
     This tries load_wave_file() load_ffmpeg_file() (for ffmpeg and avconv).
@@ -561,9 +651,13 @@ def load_audio_file(filename, sample_rate=None, num_channels=None, start=None,
     sample_rate : int, optional
         Desired sample rate of the signal [Hz], or 'None' to return the
         signal in its original rate.
-    num_channels: int, optional
-        Reduce or expand the signal to `num_channels` channels, or 'None'
-        to return the signal with its original channels.
+    num_channels : int, optional
+        Reduce or expand the signal to `num_channels` channels.
+        If 'None', return the signal with its original channels,
+        or whatever is selected by `channel`.
+    channel : int, optional
+        When reducing a signal to `num_channels` of 1, use this channel,
+        or 'None' to return the average across all channels.
     start : float, optional
         Start position [seconds].
     stop : float, optional
@@ -572,6 +666,10 @@ def load_audio_file(filename, sample_rate=None, num_channels=None, start=None,
         The data is returned with the given dtype. If 'None', it is returned
         with its original dtype, otherwise the signal gets rescaled. Integer
         dtypes use the complete value range, float dtypes the range [-1, +1].
+    replaygain_mode : {None, 'track','album'}, optional
+        Specify the ReplayGain volume-levelling mode (None to disable).
+    replaygain_preamp : float, optional
+        ReplayGain preamp volume change level (in dB).
 
     Returns
     -------
@@ -589,27 +687,21 @@ def load_audio_file(filename, sample_rate=None, num_channels=None, start=None,
     For all other audio files, this can not be guaranteed.
 
     """
-    # determine the name of the file if it is a file handle
-    try:
-        # close the file handle if it is open
-        filename.close()
-        # use the file name
-        filename = filename.name
-    except AttributeError:
-        pass
     # try reading as a wave file
     error = "All attempts to load audio file %r failed." % filename
     try:
         return load_wave_file(filename, sample_rate=sample_rate,
-                              num_channels=num_channels, start=start,
-                              stop=stop, dtype=dtype)
+                              num_channels=num_channels, channel=channel,
+                              start=start, stop=stop, dtype=dtype)
     except ValueError:
         pass
     # not a wave file (or other sample rate requested), try ffmpeg
     try:
         return load_ffmpeg_file(filename, sample_rate=sample_rate,
-                                num_channels=num_channels, start=start,
-                                stop=stop, dtype=dtype)
+                                num_channels=num_channels, channel=channel,
+                                start=start, stop=stop, dtype=dtype,
+                                replaygain_mode=replaygain_mode,
+                                replaygain_preamp=replaygain_preamp)
     except OSError as e:
         # if it's not a file not found error, raise it!
         if e.errno != errno.ENOENT:
@@ -618,9 +710,11 @@ def load_audio_file(filename, sample_rate=None, num_channels=None, start=None,
         # ffmpeg is not present, try avconv
         try:
             return load_ffmpeg_file(filename, sample_rate=sample_rate,
-                                    num_channels=num_channels, start=start,
-                                    stop=stop, dtype=dtype,
-                                    cmd_decode='avconv', cmd_probe='avprobe')
+                                    num_channels=num_channels, channel=channel,
+                                    start=start, stop=stop, dtype=dtype,
+                                    cmd_decode='avconv', cmd_probe='avprobe',
+                                    replaygain_mode=replaygain_mode,
+                                    replaygain_preamp=replaygain_preamp)
         except OSError as e:
             if e.errno == errno.ENOENT:
                 error += " Try installing ffmpeg (or avconv on Ubuntu Linux)."