+++ madmom/io/midi.py
@@ -9,6 +9,7 @@ from __future__ import absolute_import, division, print_function
 
 import numpy as np
 import mido
+import warnings
 
 DEFAULT_TEMPO = 500000  # microseconds per quarter note (i.e. 120 bpm in 4/4)
 DEFAULT_TICKS_PER_BEAT = 480  # ticks per quarter note
@@ -123,7 +124,7 @@ class MIDIFile(mido.MidiFile):
 
         - 'absolute', 'abs', 'a': use absolute timing.
         - 'relative', 'rel', 'r': use relative timing, i.e. delta to
-        previous message.
+          previous message.
 
     Examples
     --------
@@ -209,12 +210,9 @@ class MIDIFile(mido.MidiFile):
 
     """
 
-    UNIT = 'seconds'
-    TIMING = 'absolute'
-
     def __init__(self, filename=None, file_format=0,
-                 ticks_per_beat=DEFAULT_TICKS_PER_BEAT, unit=UNIT,
-                 timing=TIMING, **kwargs):
+                 ticks_per_beat=DEFAULT_TICKS_PER_BEAT, unit='seconds',
+                 timing='absolute', **kwargs):
         # instantiate a MIDIFile
         super(MIDIFile, self).__init__(filename=filename, type=file_format,
                                        ticks_per_beat=ticks_per_beat, **kwargs)
@@ -272,7 +270,7 @@ class MIDIFile(mido.MidiFile):
     @property
     def tempi(self):
         """
-        Tempi (mircoseconds per quarter note) of the MIDI file.
+        Tempi (microseconds per quarter note) of the MIDI file.
 
         Returns
         -------
@@ -294,7 +292,7 @@ class MIDIFile(mido.MidiFile):
         if not tempi or tempi[0][0] > 0:
             tempi.insert(0, (0, DEFAULT_TEMPO))
         # tempo is given in microseconds per quarter note
-        # TODO: add otption to return in BPM
+        # TODO: add option to return in BPM
         return np.asarray(tempi, np.float)
 
     @property
@@ -336,7 +334,7 @@ class MIDIFile(mido.MidiFile):
             Array with notes (onset time, pitch, duration, velocity, channel).
 
         """
-        # list for all notes
+        # lists to collect notes and sustain messages
         notes = []
         # dictionary for storing the last onset time and velocity for each
         # individual note (i.e. same pitch and channel)
@@ -352,17 +350,17 @@ class MIDIFile(mido.MidiFile):
             # use only note on or note off events
             note_on = msg.type == 'note_on'
             note_off = msg.type == 'note_off'
+            if not (note_on or note_off):
+                continue
             # hash sounding note
-            if note_on or note_off:
-                note = note_hash(msg.channel, msg.note)
-            # if it's a note on event with a velocity > 0,
+            note = note_hash(msg.channel, msg.note)
+            # start note if it's a 'note on' event with velocity > 0
             if note_on and msg.velocity > 0:
                 # save the onset time and velocity
                 sounding_notes[note] = (msg.time, msg.velocity)
-            # if it's a note off or a note on event with a velocity of 0,
+            # end note if it's a 'note off' event or 'note on' with velocity 0
             elif note_off or (note_on and msg.velocity == 0):
                 if note not in sounding_notes:
-                    import warnings
                     warnings.warn('ignoring MIDI message %s' % msg)
                     continue
                 # append the note to the list
@@ -374,6 +372,87 @@ class MIDIFile(mido.MidiFile):
 
         # sort the notes and convert to numpy array
         return np.asarray(sorted(notes), dtype=np.float)
+
+    @property
+    def sustain_messages(self):
+        """
+        Sustain messages of the MIDI file.
+
+        Returns
+        -------
+        sustain_messages : list
+            List with MIDI sustain messages.
+
+        Notes
+        -----
+        If the last sustain message is a 'sustain on' message (i.e. it has a
+        value >= 64), an artificial sustain message with a value of 0 and the
+        timing of the last MIDI message is appended to the list.
+
+        """
+        sustain_msgs = []
+        last_msg_time = None
+        for msg in self:
+            last_msg_time = msg.time
+            # keep track of sustain information
+            if msg.type == 'control_change' and msg.control == 64:
+                sustain_msgs.append(msg)
+        # if the last sustain message is 'sustain on', append a fake sustain
+        # message to end sustain with the last note
+        if sustain_msgs and sustain_msgs[-1].value >= 64:
+            msg = sustain_msgs[-1].copy()
+            msg.time = last_msg_time
+            msg.value = 0
+            sustain_msgs.append(msg)
+        return sustain_msgs
+
+    @property
+    def sustained_notes(self):
+        """
+        Notes of the MIDI file with applied sustain information.
+
+        Returns
+        -------
+        notes : numpy array
+            Array with notes (onset time, pitch, duration, velocity, channel).
+
+        """
+        notes = np.copy(self.notes)
+        # apply sustain information
+        # keep track of sustain start times (channel = key)
+        sustain_starts = {}
+        note_offsets = notes[:, 0] + notes[:, 2]
+        for msg in self.sustain_messages:
+            # remember sustain start
+            if msg.value >= 64:
+                if msg.channel in sustain_starts:
+                    # sustain is ON already, ignoring
+                    continue
+                sustain_starts[msg.channel] = msg.time
+            # expand all notes in this channel until sustain end
+            else:
+                if msg.channel not in sustain_starts:
+                    # sustain is OFF already, ignoring
+                    continue
+                # end all notes with i) offsets between sustain start and end
+                sustained = np.logical_and(
+                    note_offsets >= sustain_starts[msg.channel],
+                    note_offsets <= msg.time)
+                # and ii) same channel
+                sustained = np.logical_and(sustained,
+                                           notes[:, 4] == msg.channel)
+                # update duration of notes (sustain end time - onset time)
+                notes[sustained, 2] = msg.time - notes[sustained, 0]
+                # remove sustain start time for this channel
+                del sustain_starts[msg.channel]
+        # end all notes latest when next note (of same pitch) starts
+        for pitch in np.unique(notes[:, 1]):
+            note_idx = np.nonzero(notes[:, 1] == pitch)[0]
+            max_duration = np.diff(notes[note_idx, 0])
+            notes[note_idx[:-1], 2] = np.minimum(notes[note_idx[:-1], 2],
+                                                 max_duration)
+        # finally return notes
+        return notes
 
     @classmethod
     def from_notes(cls, notes, unit='seconds', tempo=DEFAULT_TEMPO,
@@ -439,13 +518,12 @@ class MIDIFile(mido.MidiFile):
         for note in notes:
             try:
                 onset, pitch, duration, velocity, channel = note
-                channel = int(channel)
-                velocity = int(velocity)
             except ValueError:
                 onset, pitch, duration, velocity = note
                 channel = 0
             pitch = int(pitch)
             velocity = int(velocity)
+            channel = int(channel)
             offset = onset + duration
             # create MIDI messages
             onset = second2tick(onset, ticks_per_beat, tempo)
@@ -479,14 +557,16 @@ class MIDIFile(mido.MidiFile):
             self._save(f)
 
 
-def load_midi(filename):
+def load_midi(filename, sustain=False):
     """
     Load notes from a MIDI file.
 
     Parameters
     ----------
-    filename: str
+    filename : str
         MIDI file.
+    sustain : bool, optional
+        Apply sustain information to the notes.
 
     Returns
     -------
@@ -494,6 +574,8 @@ def load_midi(filename):
         Notes ('onset time' 'note number' 'duration' 'velocity' 'channel')
 
     """
+    if sustain:
+        return MIDIFile(filename).sustained_notes
     return MIDIFile(filename).notes