+++ madmom/processors.py
@@ -64,12 +64,6 @@ class Processor(object):
             except TypeError:
                 # Python 2 doesn't have/need the encoding
                 obj = pickle.load(f)
-        # warn if the unpickled Processor is of other type
-        if obj.__class__ is not cls:
-            import warnings
-            warnings.warn("Expected Processor of class '%s' but loaded "
-                          "Processor is of class '%s', processing anyways." %
-                          (cls.__name__, obj.__class__.__name__))
         return obj
 
     def dump(self, outfile):
@@ -760,6 +754,11 @@ class BufferProcessor(Processor):
         self.init = init
         self.data = init
 
+    @property
+    def buffer_length(self):
+        """Length of the buffer (time steps)."""
+        return self.buffer_size[0]
+
     def reset(self, init=None):
         """
         Reset BufferProcessor to its initial state.
@@ -786,6 +785,12 @@ class BufferProcessor(Processor):
         numpy array or subclass thereof
             Data with buffered context.
 
+        Notes
+        -----
+        If the length of data is the same as the buffer's length, the data of
+        the buffer is completely overwritten by new data. If it exceeds the
+        length, only the latest 'buffer_length' items of data are used.
+
         """
         # expected minimum number of dimensions
         ndmin = len(self.buffer_size)
@@ -794,9 +799,14 @@ class BufferProcessor(Processor):
             data = np.array(data, copy=False, subok=True, ndmin=ndmin)
         # length of the data
         data_length = len(data)
-        # remove `data_length` from buffer at the beginning and append new data
-        self.data = np.roll(self.data, -data_length, axis=0)
-        self.data[-data_length:] = data
+        # if length of data exceeds buffer length simply replace buffer data
+        if data_length >= self.buffer_length:
+            self.data = data[-self.buffer_length:]
+        else:
+            # roll buffer by `data_length`, i.e. move data to the 'left'
+            self.data = np.roll(self.data, -data_length, axis=0)
+            # overwrite 'right' part with new data
+            self.data[-data_length:] = data
         # return the complete buffer
         return self.data
 
@@ -832,7 +842,7 @@ def process_online(processor, infile, outfile, **kwargs):
         Processor to be processed.
     infile : str or file handle, optional
         Input file (handle). If none is given, the stream present at the
-        system's audio inpup is used. Additional keyword arguments can be used
+        system's audio input is used. Additional keyword arguments can be used
         to influence the frame size and hop size.
     outfile : str or file handle
         Output file (handle).
@@ -851,7 +861,7 @@ def process_online(processor, infile, outfile, **kwargs):
     # set default values
     kwargs['sample_rate'] = kwargs.get('sample_rate', 44100)
     kwargs['num_channels'] = kwargs.get('num_channels', 1)
-    # if no iput file is given, create a Stream with the given arguments
+    # if no input file is given, create a Stream with the given arguments
     if infile is None:
         # open a stream and start if not running already
         stream = Stream(**kwargs)